Use uint32 to work with network id (#2479)
Int8 is too short and forces reuse of the same ids for testnets.
Closes: https://github.com/spacemeshos/go-spacemesh/issues/2113

- NetworkID is stored internally as uint32 instead of int8.
- handshake data will use uint32 instead of int32 for consistency

Existing unit and system tests.

- [x] Explain motivation or link existing issue(s)
- [x] Test changes and document test plan
- [ ] Update documentation as needed

- [x] This PR does not require configuration changes (e.g., environment variables, GitHub secrets, VM resources)
- [x] This PR does not affect public APIs
- [x] This PR does not rely on a new version of external services (PoET, elasticsearch, etc.)
- [x] This PR does not make changes to log messages (which monitoring infrastructure may rely on)

+++ api/grpcserver/mesh_service.go
@@ -19,7 +19,7 @@ type MeshService struct {
 	Mempool           api.MempoolAPI
 	GenTime           api.GenesisTimeAPI
 	LayersPerEpoch    int
-	NetworkID         int8
+	NetworkID         uint32
 	LayerDurationSec  int
 	LayerAvgSize      int
 	TxsPerBlock       int
@@ -36,7 +36,7 @@ func (s MeshService) RegisterService(server *Server) {
 // NewMeshService creates a new service using config data
 func NewMeshService(
 	tx api.TxAPI, mempool api.MempoolAPI, genTime api.GenesisTimeAPI,
-	layersPerEpoch int, networkID int8, layerDurationSec int,
+	layersPerEpoch int, networkID uint32, layerDurationSec int,
 	layerAvgSize int, txsPerBlock int) *MeshService {
 	return &MeshService{
 		Mesh:             tx,
